simplified backup file creation
