+++ mRemoteNGTests/mRemoteNGTests.csproj
@@ -112,8 +112,8 @@
     <Compile Include="App\UpdaterTests.cs" />
     <Compile Include="BinaryFileTests.cs" />
     <Compile Include="Config\Connections\Multiuser\ConnectionsUpdateAvailableEventArgsTests.cs" />
+    <Compile Include="Config\DataProviders\FileBackupCreatorTests.cs" />
     <Compile Include="Config\DataProviders\FileDataProviderTests.cs" />
-    <Compile Include="Config\DataProviders\FileDataProviderWithBackupTests.cs" />
     <Compile Include="Config\DataProviders\FileDataProviderWithRollingBackupTests.cs" />
     <Compile Include="Config\Serializers\DataTableDeserializerTests.cs" />
     <Compile Include="Config\CredentialHarvesterTests.cs" />
