--- mRemoteV1/Config/DataProviders/FileDataProviderWithBackup.cs
@@ -1,43 +1,0 @@
-﻿using System;
-using System.IO;
-using mRemoteNG.App;
-
-namespace mRemoteNG.Config.DataProviders
-{
-    public class FileDataProviderWithBackup : FileDataProvider
-    {
-        protected string BackupFileSuffix = ".backup";
-
-        public FileDataProviderWithBackup(string filePath) : base(filePath)
-        {
-        }
-
-        public override void Save(string content)
-        {
-            CreateBackup();
-            base.Save(content);
-        }
-
-        protected virtual void CreateBackup()
-        {
-            var backupFileName = FilePath + BackupFileSuffix;
-            try
-            {
-                if (WeDontNeedToBackup())
-                    return;
-
-                File.Copy(FilePath, backupFileName, true);
-            }
-            catch (Exception ex)
-            {
-                Runtime.MessageCollector.AddExceptionStackTrace($"Failed to create backup of file {FilePath}", ex);
-                throw;
-            }
-        }
-
-        private bool WeDontNeedToBackup()
-        {
-            return !File.Exists(FilePath);
-        }
-    }
-}
